+++ src/test/html-transform_test.ts
@@ -44,6 +44,9 @@ function replaceGiantScripts(html: string): string {
 }
 
 suite('htmlTransform', () => {
+  const fixtureRoot =
+      path.join(__dirname, '..', '..', 'test-fixtures', 'npm-modules');
+
   test('minifies html', () => {
     const input = `
       <html>
@@ -110,8 +113,6 @@ suite('htmlTransform', () => {
   });
 
   test('rewrites bare module specifiers to paths', () => {
-    const fixtureRoot =
-        path.join(__dirname, '..', '..', 'test-fixtures', 'npm-modules');
     const filePath = path.join(fixtureRoot, 'foo.html');
 
     const input = `
@@ -149,7 +150,7 @@ suite('htmlTransform', () => {
       assertEqualIgnoringWhitespace(
           htmlTransform(input, {
             js: {
-              transformEsModulesToAmd: true,
+              transformModulesToAmd: true,
             }
           }),
           expected);
@@ -177,7 +178,13 @@ suite('htmlTransform', () => {
         </body></html>`;
 
       assertEqualIgnoringWhitespace(
-          htmlTransform(input, {js: {transformEsModulesToAmd: true}}),
+          htmlTransform(input, {
+            js: {
+              transformModulesToAmd: true,
+              filePath: path.join(fixtureRoot, 'foo.html'),
+              rootDir: fixtureRoot,
+            },
+          }),
           expected);
     });
 
@@ -200,15 +207,17 @@ suite('htmlTransform', () => {
       </body></html>`;
 
       assertEqualIgnoringWhitespace(
-          htmlTransform(input, {js: {transformEsModulesToAmd: true}}),
+          htmlTransform(input, {
+            js: {
+              transformModulesToAmd: true,
+              filePath: path.join(fixtureRoot, 'foo.html'),
+              rootDir: fixtureRoot,
+            }
+          }),
           expected);
     });
 
     test('resolves names and does AMD transform', () => {
-      const fixtureRoot =
-          path.join(__dirname, '..', '..', 'test-fixtures', 'npm-modules');
-      const filePath = path.join(fixtureRoot, 'foo.html');
-
       const input = `
         <html><head></head><body>
           <script type="module">import { dep1 } from 'dep1';</script>
@@ -223,9 +232,10 @@ suite('htmlTransform', () => {
       assertEqualIgnoringWhitespace(
           htmlTransform(input, {
             js: {
-              transformEsModulesToAmd: true,
+              transformModulesToAmd: true,
               moduleResolution: 'node',
-              filePath,
+              filePath: path.join(fixtureRoot, 'foo.html'),
+              rootDir: fixtureRoot,
             }
           }),
           expected);
@@ -246,7 +256,7 @@ suite('htmlTransform', () => {
           htmlTransform(input, {
             js: {
               compileToEs5: true,
-              transformEsModulesToAmd: true,
+              transformModulesToAmd: true,
             }
           }),
           expected);
@@ -266,7 +276,7 @@ suite('htmlTransform', () => {
         </body></html>`;
 
       assertEqualIgnoringWhitespace(
-          htmlTransform(input, {js: {transformEsModulesToAmd: true}}), input);
+          htmlTransform(input, {js: {transformModulesToAmd: true}}), input);
     });
 
     test('adds AMD loader to entry point before first module', () => {
@@ -291,7 +301,7 @@ suite('htmlTransform', () => {
       const result = htmlTransform(input, {
         injectAmdLoader: true,
         js: {
-          transformEsModulesToAmd: true,
+          transformModulesToAmd: true,
         },
       });
       assertEqualIgnoringWhitespace(replaceGiantScripts(result), expected);
@@ -307,7 +317,7 @@ suite('htmlTransform', () => {
       const result = htmlTransform(input, {
         injectAmdLoader: true,
         js: {
-          transformEsModulesToAmd: true,
+          transformModulesToAmd: true,
         },
       });
       assertEqualIgnoringWhitespace(result, input);
@@ -339,7 +349,7 @@ suite('htmlTransform', () => {
       const result = htmlTransform(input, {
         injectAmdLoader: true,
         js: {
-          transformEsModulesToAmd: true,
+          transformModulesToAmd: true,
         },
       });
       assertEqualIgnoringWhitespace(replaceGiantScripts(result), expected);